Fix share tracker ui
